backend/requirements.txt
mysql-connector-python
fastapi[standard]
datetime